Add Point drawing

+++ lab_01/lab_01/lab_01/Point.cs
@@ -8,10 +8,10 @@ namespace lab_01
 {
     class Point
     {
-        public double x;
-        public double y;
+        public float x;
+        public float y;
 
-        public Point(double x, double y)
+        public Point(float x, float y)
         {
             this.x = x;
             this.y = y;
